requirements-dev.txt
docx-mailmerge # TNGPS patch-1 https://github.com/TNGPS/docx-mailmerge/tree/patch-1
qrcode[pil]